requirements.txt
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
bokeh==2.4.3
boto3==1.21.21
botocore==1.24.21
branca==0.5.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.1.0
coiled==0.2.15
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
dask==2022.7.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.3
folium==0.12.1.post1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
HeapDict==1.0.1
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsondiff==2.0.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
locket==1.0.0
markdown2==2.4.13
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
msgpack==1.0.4
multidict==6.0.2
multimethod==1.8
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.11
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.1
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.23.0
protobuf==3.20.0
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.5
pyxlsb==1.0.9
PyYAML==6.0
pyzmq==23.0.0
requests==2.27.1
rich==12.5.1
rsa==4.9
s3fs==2022.5.0
s3transfer==0.5.2
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.9.2
streamlit-folium==0.6.14
streamlit-multipage==0.0.18
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tblib==1.7.0
tenacity==8.0.1
terminado==0.15.0
textual==0.1.18
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
visions==0.7.4
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.14.1
XlsxWriter==3.0.3
yarl==1.7.2
zict==2.2.0
zipp==3.8.0
